requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
deep-translator==1.11.4
Django==5.1.7
django-ckeditor-5==0.2.17
django-colorfield==0.12.0
django-debug-toolbar==5.1.0
django-jazzmin==3.0.1
django-tinymce==4.1.0
fonttools==4.57.0
googletrans==4.0.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2docx==1.6.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
kiwisolver==1.4.8
lxml==5.3.1
Markdown==3.7
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
termcolor==2.5.0
tinycss2==1.4.0
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
